Audit rounds 3-5: 72 correctness fixes + 28 new regression tests
Rounds 3-5 of the paper-grade correctness audit (AUDIT_REPORT.md), spanning:

Round 3 (paper-faithfulness):
  - matrix_entropy: per-sentence SVD, row L2 normalize, /log(d) (Wei et al. 2024)
  - tracing: sweep every layer, average over n_noise_samples=10 (Meng 2022)
  - CoE: emit Eq. 5 (CoE-R) and Eq. 7 (CoE-C) from Wang 2025
  - Two-NN: linear-regression form, F=i/N, paper-faithful tail trim
  - neural_collapse: subspace-projected NC1 (Papyan 2020 Sec. 2.3)
  - unified score_continuation helper for tokenization-boundary tasks
  - effective_rank unified on p=sigma^2/Sigma sigma^2 (Roy-Vetterli)

Round 4 (aggregator + parallel audit):
  - aggregator: regroup layer-indexed keys so per-layer columns become
    architecture-agnostic mean/std/slope/q25/q50/q75 summaries
  - cond_number: cap by numerical rank so GPT-2 no longer reports 1.4e8
  - ppl_rare: threshold only observed tokens (was +inf for all models)
  - Mahalanobis: held-out ID split (avoid in-sample leverage bias)
  - LID: seeded RNG (reproducibility)
  - betti_curve: threshold = median pairwise distance (geometry-invariant)
  - gradient_flow: regress on normalised depth
  - induction: raw accuracy diff (not divided by model-dependent baseline)
  - attention_graph: zero-total Gini guard
  - prediction_entropy: log_softmax for bf16/fp16 stability
  - bf16 .float().cpu().numpy() fixes in 4 tasks

Round 5 (previously-un-audited tasks):
  - WEAT: offset-mapping word position (was falling back to end-of-sentence)
  - ablation / edge_attribution / trajectories / attention_polysemanticity:
    seeded RNGs
  - sae_features: target_layer parsed from sae_id
  - repe_steering_effectiveness: cast task vector to hidden dtype
  - persistence_landscape: np.trapezoid with np.trapz fallback
  - prediction_alignment: use lm_head.weight (untied-head correctness)

New regression tests (28):
  - tests/test_aggregate_results.py         — depth-bias regrouping
  - tests/tasks/test_mahalanobis_task.py    — held-out split
  - tests/tasks/test_perplexity_task.py     — observed-tokens thresholds
  - tests/tasks/test_round4_fixes.py        — round-4 fixes
  - tests/tasks/test_round5_determinism.py  — determinism + repe dtype
  - tests/tasks/test_bias_task.py           — WEAT offset-mapping

All 244 tests pass (1 skipped). All 32 models patched with fixes applied.

+++ src/blme/tasks/consistency/contamination.py
@@ -69,8 +69,10 @@ class ContaminationDetectionTask(DiagnosticTask):
                 shift_logits = logits[:, :-1, :]
                 shift_labels = input_ids[:, 1:]
 
-                # Per-token log probabilities
-                log_probs = F.log_softmax(shift_logits, dim=-1)
+                # Per-token log probabilities — cast to fp32 before
+                # log_softmax so bf16 models (which otherwise return
+                # bf16 log-probs that numpy cannot convert) work.
+                log_probs = F.log_softmax(shift_logits.float(), dim=-1)
                 token_log_probs = log_probs.gather(
                     2, shift_labels.unsqueeze(-1)
                 ).squeeze(-1)  # (1, T-1)
